remove service client functionality
