+++ build.gradle
@@ -24,10 +24,10 @@ repositories {
 
 dependencies {
     //implementation 'org.iot-dsa:dslink-v2-websocket:+' //for a locally installed sdk
-    implementation 'com.github.iot-dsa-v2.sdk-dslink-java-v2:dslink-v2-websocket:0.54.0'
+    implementation 'com.github.iot-dsa-v2.sdk-dslink-java-v2:dslink-v2-websocket:+'
     implementation 'org.apache.commons:commons-lang3:3.6'
-    implementation 'com.microsoft.azure:azure-eventhubs:0.13.0'
-	implementation 'com.microsoft.azure.sdk.iot:iot-service-client:1.5.22'
+    //implementation 'com.microsoft.azure:azure-eventhubs:0.13.0'
+	//implementation 'com.microsoft.azure.sdk.iot:iot-service-client:1.5.22'
 	implementation 'com.microsoft.azure.sdk.iot:iot-device-client:1.14.1'
 	
     testImplementation 'junit:junit:+'
