+++ src/main/java/org/iot/dsa/iothub/MainNode.java
@@ -4,12 +4,15 @@ import org.iot.dsa.dslink.DSIRequester;
 import org.iot.dsa.dslink.DSLinkConnection;
 import org.iot.dsa.dslink.DSMainNode;
 import org.iot.dsa.node.DSInfo;
+import org.iot.dsa.node.DSJavaEnum;
 import org.iot.dsa.node.DSMap;
 import org.iot.dsa.node.DSString;
 import org.iot.dsa.node.DSValueType;
 import org.iot.dsa.node.action.ActionInvocation;
 import org.iot.dsa.node.action.ActionResult;
 import org.iot.dsa.node.action.DSAction;
+import org.iot.dsa.util.DSException;
+import com.microsoft.azure.sdk.iot.device.IotHubClientProtocol;
 
 /**
  * This is the root node of the link.
@@ -31,35 +34,44 @@ public class MainNode extends DSMainNode {
     @Override
     protected void declareDefaults() {
         super.declareDefaults();
-        DSAction act = new DSAction.Parameterless() {
-            @Override
-            public ActionResult invoke(DSInfo target, ActionInvocation invocation) {
-                ((MainNode) target.get()).handleAddIotHub(invocation.getParameters());
-                return null;
-            }
-        };
-        act.addParameter("Name", DSValueType.STRING, null);
-        act.addParameter("Connection String", DSValueType.STRING, null);
-        declareDefault("Add IoT Hub", act);
         declareDefault("Docs", DSString.valueOf(
                 "https://github.com/iot-dsa-v2/dslink-java-v2-iothub/blob/develop/README.md"))
                 .setTransient(true).setReadOnly(true);
+        
+        declareDefault("Add Device by Connection String", makeAddDeviceByConnStrAction());
+        //declareDefault("Add Device by DPS", makeAddDeviceByDPSAction());
     }
 
     @Override
     protected void onStarted() {
-        getLink().getUpstream().subscribe(((event, node, child, data) -> {
+        getLink().getConnection().subscribe(((event, node, child, data) -> {
             if (event.equals(DSLinkConnection.CONNECTED_EVENT)) {
-                MainNode.setRequester(getLink().getUpstream().getRequester());
+                MainNode.setRequester(getLink().getConnection().getRequester());
             }
         }));
     }
-
-    private void handleAddIotHub(DSMap parameters) {
-        String name = parameters.getString("Name");
-        String connString = parameters.getString("Connection String");
-
-        IotHubNode hub = new IotHubNode(connString);
-        add(name, hub);
+    
+    private static DSAction makeAddDeviceByConnStrAction() {
+        DSAction act = new DSAction.Parameterless() {
+            @Override
+            public ActionResult invoke(DSInfo target, ActionInvocation invocation) {
+                ((MainNode) target.get()).addDeviceByConnStr(invocation.getParameters());
+                return null;
+            }
+        };
+        act.addParameter("Connection String", DSValueType.STRING, null);
+        act.addParameter("Protocol", DSJavaEnum.valueOf(IotHubClientProtocol.MQTT), null);
+        return act;
+    }
+    
+    private void addDeviceByConnStr(DSMap parameters) {
+        String connStr = parameters.getString("Connection String"); 
+        String id = Util.getFromConnString(connStr, "DeviceId");
+        if (id == null) {
+            DSException.throwRuntime(new IllegalArgumentException("Device Connection String missing Device ID"));
+        }
+        String protocolStr = parameters.getString("Protocol");
+        IotHubClientProtocol protocol = IotHubClientProtocol.valueOf(protocolStr);
+        add(id, new LocalDeviceNode(id, protocol, connStr));
     }
 }
